+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Unreleased
+## [0.21.0] - 2025-05-25
 ### Added
 - Support for colored gridlines with the options `x_gridlines_color` and
   `y_gridlines_color`.